Create Social_Media_War_With_Menu_System.py
In this, many changed things from the first code new single-player mode is added in this code. A Start Menu is also added which asks the user to specify which mode to play single or two. And also an end menu that appears after the game ends with options to rematch Go to the start menu and quit the game.